requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
redis==5.2.1
passlib[bcrypt]
python-jose== 3.4.0
PyJWT==2.10.1
python-multipart
bcrypt==3.2.0
pytest==8.3.5
locust==2.33.2
coverage==7.8.0
httpx==0.28.1
fakeredis==2.28.0